feature: add e2e encryption for chat messages

+++ client/src/routes/+page.ts
@@ -3,9 +3,18 @@ export const ssr = false;
 import { errorStore } from '$lib/stores/errorStore';
 import { get } from 'svelte/store';
 import type { PageLoad } from './$types';
+import { getItemFromSessionStorage } from "$lib/api/utils.js";
 
 export const load: PageLoad = async () => {
-    sessionStorage.clear();
+    const isSameSession =
+        !!getItemFromSessionStorage("activeRoomId") ||
+        !!getItemFromSessionStorage("token") ||
+        !!getItemFromSessionStorage("encryptionKey");
+
+    if (isSameSession) {
+        sessionStorage.clear();
+        return { errorMessage: "You navigated to the home page from the chat room. The chat room is now closed." }
+    }
 
     let errorReason = get(errorStore);
     errorStore.set(null);
@@ -22,6 +31,7 @@ export const load: PageLoad = async () => {
         "user-not-exists": "Unexpected error: you were not found in this room.",
         "message-too-large": "Unexpected error: the size of the message you sent was too large.",
         "error": "An unexpected server error occurred.",
+        "invalid-session": "Unexpected error: encryption key is missing, please try again."
     };
 
     return {
